Fix due to changes to Cucumber::ScenarioOutline.
As of #fcc418, a ScenarioOutline#line is a private method and we should use
Cucumber::Ast::Location to get the line number instead.

This ensures everything keeps working, although perhaps it would be a
good idea to refactor the tests around this in a way that would catch
changes to the dependency on Cucumber.

As a general rule, we shouldn't mock what we don't own as it opens up
the possibility of masking breaking changes on dependencies.

I could look at submitting a pull request with some suggestions for a
refactor if you like?